move tab component to own file
